Replace zstd by zip, remove empty BatchNorm params.
Apparently the zstd that is included in current Linux distributions
can't necessarily decompress files produced by the current version. So
switch back to the true and tested ZIP and link to a new zipped weights
file.

Use the opportunity to remove the empty last BatchNorm weight layer
produced by NVIDIA-Caffe. Remove it from the specification and adapt
the weight reader accordingly.