Randomization of the root move proportional to visits.
To improve randomness further in self-play, we optionally allow
replacing the best (first) move with the most visits by another move.
The replacement move is selected proportionally to its visit count. This
is equivalent to a softmax with temperature=1. Not doing the switch is
equivalent to softmax with temperature=0.

This is done for the first few moves in the game, where the amount of
moves is a configurable parameter.