Error out when we are given arguments.
Currently we eat and ignore any non-option arguments on the command
line. Some users seem to be of the impression that GTP commands can
somehow be specified there and are then surprised that they are not
taken into account.

Eat all remaining arguments as positional options, and error out
if any are detected.

Fixes Issue #4.