Add "dump_training" command.
Add support for dumping the training data in a simple text based format
at the end of the game. This should be suitable and easy to parse for
most Deep Learning frameworks. Dumping the data directly also avoids the
intermediate step of converting an SGF back to the correct input
bitplanes.

It's a bit verbose, so I'm banking on this compressing *very* well. We can
revisit if that turns out to be a problem.