Allow dumping SGF files out as training data.
Add a function that reads an SGF file, parses the games, and then dumps
a fraction of the positions to a text file, suitable for use in Deep
Learning frameworks.

We'll probably want to add some functionality to split it up in chunks
and compress them, or else it's going to eat disk space. But for
testing, this is good enough for now.