Add autogtp player, fix printsgf/training.
Add the autogtp program to play self-play matches. It will run a
match between two engines at the appropriate settings (with
randomization enabled), and at the end dump the SGF and the training
data to a unique filename.

Make the printsgf command properly conform to the GTP protocol, and make
sure dump_training appends rather than replaces the destination file.