Don't pass out a loss if the natural move is passing.
If the search returns a pass move, it was possible for us to check that
the opponent had passed, and hence us passing ends the game, see that it
loses, yet still play a pass because the search move was not overruled.

Change the order of the checks so a pass move from ourselves has
priority.