Don't override an engine pass with an eye-fill.
When using kgs-genmove_cleanup, we try to play out the position as much
as possible and remove dead groups. We do this by overriding the engine
and selecting the second best move if the first move is a pass. However,
this means that we eventually will pick moves that fill eyes. So stop if
we've arrived at this point.

Note that the AlphaGo paper specifically states that Zero doesn't have
this knowledge. Our engine doesn't either - it's only used when we are
overruling its move selection for scoring a position.