Simplify building, especially macOS.
Clean up the Makefile a bit so the libraries and includes for each
platform are grouped together, making it easier to find what to comment
out and what not.

Include the OpenCL C++ header by default in the source tree, saving a
step and making more sure it's up to date.

Update the README with example steps for macOS.