Don't update nodes when not expanded.
There's a corner case right now where upon hitting the tree size limit,
the node expansion and update won't run, but due to the use of virtual
losses, the score for the node will drop as it did get a visit added. If
the search runs for a while with the tree size limit hit, all scores
for black will go to 0%.

Split off the virtual loss/win code into explicit calls, and don't let
the node expansion set the evaluation result, but pass it into the
caller, who can then call the update (if the result was valid)
explicitly.