Fix virtual loss induced races.
Due to the use of virtual loss, it is possible for a node to appear to
have been visited multiple times before, and then, when calculating the
score, suddenly have no visits, causing division by zero problems. Cache
the visits and score before making this decision and handle the
first-play-urgency at the same time.

Similarly, it is possible to update the hashtable with a visit number
that contains virtual losses, and due to the higher number of visits
taking priority, get this result to "stick". Solve this by storing the
virtual losses in a separate counter, so we can use the untainted visit
count to see the number of node updates. Virtual loss considerations can
then be limited to only to score calculation, which is used for
choosing the next move to search.