Find OpenBLAS specifically in CMake, fix linking.
We need OpenBLAS specific functions, so indicate that we need this
specific vendor in the CMake config. Also make sure we pick up the
OpenBLAS headers, and not the generic BLAS headers.

Boost needs to be told which sub-library is specifically requested, and
there was a captilization difference between the current code and what
the manuals use.