Allow AutoGTP to check Leela Zero's version.
The server can ask for an AutoGTP update, but not for a Leela Zero
update directly. So implement a version check in AutoGTP. If necessary
we can then force a Leela Zero update by bumping the AutoGTP version.

This also bumps both versions, so we can request the fix for issue #109.