Remove some code relics, handle user input suicide.
We don't update the ko-hash for suicide moves, because we assume we
don't care about repetition detection for cases where we generate them.
Right now, Leela doesn't generate suicide moves at all, but it does
still allows the user to enter them. So we should make sure we calculate
the ko-hash correctly.

Reduce move history to 2 entries because that's all we use.

Fixes issue #446.