Make fixSgf more robust.
Greedy .* is dangerous, it will gobble up the entire sgf file if
there is a space later on in it.
Also add the sgf keywords PB and PW to the match.

There is a similar problem in the if(resignation) block with
greedy .*. It only works because RE[] is the last property on the
header line of the sgf file. I didn't touch that because I think
that code is going away soon.