Replace TTable by NNCache.
NNCache doubles as a sort of TTable now by using board.get_hash()
instead of hashing the input planes. Experiments suggest that
reconstructing the tree by repeatedly hitting NNCache may be stronger
in real games than only getting the results from the TTable, even if it
may be weaker at a fixed number of playouts. Given that this is good
code simplification, we'll get rid of the TTable for now.

Pull request #662.