Trim the search tree during the search.
In the master, do a check whether we are 3/4ths through our memory
budget. If so, attempt to do a depth first pruning of the search tree.

We try to get a lock on a node and check if there are any other CPUs
searching inside it. If so, we give up. If not, we prune every child
that has less than a given number of visits. These can be re-expanded by
the search later on. We find the right threshold by just doubling it
until the number of nodes in the tree drops below half our memory budget.

With a lot of threads active, we might not necessarily be able to prune
enough, especially with low maximum tree sizes, so we will still abort
if we reach the full memory budget.

This reworks the virtual loss code to accurately reflect the number of
active CPUs in a subtree.