Disable FPU reductions at a noised root.
The Dirichlet noise that is added to self-play games was originally
tuned for a regular UCT search in the AGZ paper. With FPU reduction,
the possibility to discover a noised move is greatly reduced. So
we should not use FPU reduction at the place where we add noise, i.e. at
the root.