Batch norm beta support.
This adds support for "centering" after the BatchNorm layers, by
storing the relevant weights in the layer meant for the biases after 
the convolution weights, and adjusting it for the current batchnorm 
weights. Note that Leela Zero will do an opposite adjustment when 
loading the network and integrate the convolution biases (aka the 
centering) into the adjusted batchnorm mean.

Pull request #1099.