Alternate Thompson sampling implementation.
Put all tweaks in the calculation of the prior. We reduce the estimated
mean value by transforming the policy prior with a formula similar to
FPU reductions, and scale the alpha/beta values of the beta distribution
with tunables to affect both how quickly to adopt the distribution, as
how strongly to take into account the priors we set.