Change defaults for fast tuner.
Given the nature of the computations we do, there are little reasons to
believe that vector sizes of 1 in the matrix calculation are ever going
to be much - if any - better than 2 or 4, and indeed on the vast
majority of cards 1 is simply worse. So we can remove this option from
the fast tuning presets.

Similar reasoning applies to caching in local memory, which is a gain,
or at the very least never a lot worse, on modern GPU designs.

Browsing through the CLBlast results shows the only cards where small
vectors or no cache are ever listed are older, in some cases even
pre-GCN AMD designs (for which AMD no longer makes OpenCL drivers...).
Even so, they won't necessarily be worse with wide vectors or caching.

Removing these combinations makes the fast tuner much faster. Use some
of the gained speed to try a few more combinations that do help modern
hardware, while still being about twice as fast to tune overall.

The odd card that did need size 1 vectors or no local memory caching can
still use the full tuner to get optimal performance.

See issue #1273 for more benchmarking data in support of this change.