Don't exit early when playouts are limited.
When running with fixed playouts per move, even if we will not change
our selection of the best move, it is still beneficial to continue
thinking and expand our tree to benefit the search of the next move,
which will now accumulate more visits in total. (This reasoning does
not apply when running with fixed visits because the next search will
simply be shorter)

So disable early exits in the time management code for this case.

This is a trivial fix if our representation of "unlimited playouts" is
consistent, so fix that up first.

See also:
https://github.com/gcp/leela-zero/issues/1127#issuecomment-385343236