Support Facebook's ELF Open Go networks.
Although the weights for those networks can be converted to be Leela
Zero compatible, their value output is for black instead of the side to
move. As this can't be fixed up in the network itself, add support for
a new network format version to cover those.

Note that the training code isn't changed, so training such a network
further is currently not possible.

See issue #1311.

Pull request #1312.