We work with only OpenCL 1.1 support.
The problem of verifying that OpenCL 1.1 support works was always that
drivers that are stuck on 1.1 tend to be so outdated that they are just
broken and fail for others reasons.

However, I'm happy to report that Mesa 18.0.3 together with the open
source amdgpu drivers, which only claim OpenCL 1.1 support, finally
work correctly with Leela Zero. On a Polaris based card, they even
seem faster than the proprietary AMDGPU-PRO drivers.

And this confirms that only OpenCL 1.1 is required.