Look for symmetry on NNCache lookup.
* Look for symmetrical position in cache.
* Disable NNCache symmetry in self-play.

To increase randomness from rotational assymetry.

* Only check symmetry in opening. Refactor TimeControl.

Only check for symmetries in the NNCache when we are in the 
opening (fast moving zone). Refactor TimeControl to take the 
boardsize out.

* Change bench to assymetric position.

Avoids rotation symmetry speedups, they are not typical.

* Rename rotation to symmetry, limit to early opening.

Be consistent and don't call symmetries rotations. Limit the symmetry
lookups to until halfway the opening (which is the first 30 moves on
19 x 19).

Based on pull request #1275, but without keeping the rotation array in
every board instance.

Pull request #1421.