Fix up lz-analyse formats.
We preferably store the analysis in the original format for sorting, and
only do the conversion for display at display time.

Don't use the completely meaningless tag "N" for move policy prior.
We'll use "prior" instead.

Winrate is currently output in 1/100th of a percent, so we'll use the
same format for priors. I'm not sure why winrate is not just using floats,
but I assume GUIs now already rely on this, and it might avoid some weird
bugs related to locale.