Include Eigen as BLAS replacement.
This adds Eigen as a default matrix/vector library via a submodule.
This has a load of advantages:

* It can be used as a replacement for a cBLAS library when it is
  not available, cannot be found, or is outdated compared to the
  compiler or CPU.
* Because Eigen is header only, it significantly eases the build
  prequisite requirements.
* The Eigen code paths are much more readable from a mathematical
  perspective.
* Eigen can optimize more heavily for known matrix sizes. The
  current code doesn't yet take advantage of this, though.

The downsides:

* Eigen might be a bit slower than other BLAS libraries. (Nevertheless,
  on my system it is faster than OpenBLAS)
* Binaries built with Eigen are optimized for the CPU it was compiled
  on and don't port as well to other CPUs. So you need seperate binaries
  for wider client support.

* Default Eigen in CMake, add tests.

Default the Eigen library in CMake, as it's the fastest for most
contemporary CPUs and configurations, and the easiest to build.

We can optionally use BLAS by adding the USE_BLAS define, and
will try to locate BLAS/OpenBLAS if so. This is useful for
binaries for distribution such as our releases or distros.

Split all tests to cover both Eigen and BLAS.

Update build instructions to remove BLAS as a dependency, use CMake on
all Unixy platforms, and use HTTPS.

Pull request #1692.