Add side to move in lz-analyze command.
See discussion in pull request #1642.

This adds an optional side to move in the lz-analyze command, instead
of only a posting interval. This makes the format more consistent with
all other GTP commands.

We check the amount and format of the arguments so we are backwards
compatible with GUIs that send the old format, i.e. with only a posting
interval.

Pull request #1872.