Autoselect half mode for fp16 compute.
Rework the Network initialization to pull out the OpenCL benchmarking
for precision autodetection. Add support to ForwardPipe to report
whether it is necessary to run the benchmarks. If the answer is no, and
fp16 compute works, we assume that's what we want.

This avoids the benchmarking overhead on modern AMD cards and probably
on the latest ones from NVIDIA too.

Pull request #1873.