Add a feature to exclude moves from the search.
The new GTP command "avoid" adds moves to the list of moves to avoid.
Syntax is "avoid <color> <coords> <movenum>".
The specified coords will not be considered for this color starting from
the current move until the specified move number. A "+" in front of the
move number makes it relative to the current move number.

Example: "avoid b q16 +1" forbids Black to play q16 at the next move.