Refactor out FastBoardSerializer from FastBoard and add tests (#1093)
Moving the logic related to board serialization to a separate class helps
to simplify FastBoard and allows that logic to be separately testable.
Private methods like get_columns and starpoints can be moved to
FastBoardSerializer because they don't rely on board internals.

Add the FastBoardSerializer to vcxproj files so the project will
build in Visual Studio on Windows.

Add hash, ko_hash, get_prisoners, isEye tests.
Tests for count_pliberties and is_suicide currently give accurate
results for FullBoard, but not FastBoard. This will be fixed if
the update methods can move from FullBoard into FastBaord.

Remove c_invert from FastBoard because it is not used.