#1093 Add tests for FullBoard
Move update_board to FastBoard so that the state of the board will update correctly.
Remove set_state methods in FastBoard, since update_board should be used now instead.
FullBoard behavior is now consistent with FastBoard, except that it adds management of the Zobrist hash.
The results in FastBoard tests now match those in FullBoard.