Cancel pending network evaluation when search should stop. (#2508)
When we have a lot of threads it can take a while to finish the net
evaluation.  To enhance responsiveness, this patch implements a way to
signal Network to cancel pending net evaluations (if it can).  When it
cancels a net evaluation, it throws a NetworkHaltException and
UCTSearch::play_simulaation gracefully handles it.

This helps improving responsiveness (especially when running lz-analyze)
on machines with many GPUs running 100+ threads.